rename directive module to directives
