+++ libs/datamodel/core/src/validator/directives/default.rs
@@ -1,5 +1,5 @@
+use super::{Args, DirectiveValidator};
 use crate::error::DatamodelError;
-use crate::validator::directive::{Args, DirectiveValidator};
 use crate::{ast, dml, ValueGenerator};
 use prisma_value::PrismaValue;
