+++ libs/datamodel/core/src/validator/directives/relation.rs
@@ -1,7 +1,7 @@
+use super::{Args, DirectiveValidator};
 use crate::common::names::DefaultNames;
 use crate::common::value_validator::ValueListValidator;
 use crate::error::DatamodelError;
-use crate::validator::directive::{Args, DirectiveValidator};
 use crate::{ast, dml};
 
 /// Prismas builtin `@relation` directive.
