+++ libs/datamodel/core/src/validator/mod.rs
@@ -1,3 +1,4 @@
+mod directives;
+
 pub mod ast_to_dml;
-mod directive;
 pub mod dml_to_ast;
